Fix for photos being added with empty captions
The tutorial had originally used a field name of "text" for the todo
items, but within this repository we have been using the name
"caption" instead (as being more appropriate for photos).

But recent code was added, (copied from the tutorial), trying to set a
"text" field which doesn't exist, (so that photos were being added
with empty captions).

This commit fixes that bug by using "caption" correctly rather than
"text".